1、新增通道管理代码 2、增加切换线路相关接口 3、切换spring版本

+++ joggle-server/src/main/java/com/wuweibi/bullet/entity/Device.java
@@ -46,7 +46,7 @@ public class Device extends Model<Device> {
 	 * 服务通道ID
 	 */
 	@TableField("server_tunnel_id")
-	private Long serverTunnelId;
+	private Integer serverTunnelId;
 
 	@ApiModelProperty("设备秘钥")
 	@TableField("device_secret")
